+++ tests/Pest.php
@@ -4,4 +4,5 @@ declare(strict_types=1);
 
 use DigitalDevLx\LogHole\Tests\TestCase;
 
-uses(TestCase::class)->in(__DIR__);
+uses(TestCase::class)->in('Unit', 'Feature');
+uses(TestCase::class)->in('ArchTest.php');